Add some additional Cleanup in Teardown/Verify
If one of the gcloud commands fail for some reason (i.e. flakey IAM
propagation or something), it's easy to for the repository or pull
requests to be left open.

The Cleanup method is called at the end of the test, regardless if the
test failed for some reason. Especially for the teardown step, it's nice
to not leave repositories hanging around.